Cleanup Makefile and Dockerfile  (#47)
## Makefile

By looking at our Makefiles we can identify several topics:

- building
- testing
- containerization
- help
- sanity checks (e.g. linting and formatting)
- ...

I've isolate each of those topics into their own makefiles under `./make` and and tried to reduce the number of common dependencies to a minimum. For example, if your Makefile needs to know something about make the git version, just include

```Makefile
include ./make/git.mk
```
**NOTE:** You can include each `./make/*.mk` as many times as you want. They all have include guards and are only included once.

## Dockerfile

I wanted for quay.io or docker hub to be able to take and build our code as easily as possible and that means we need to have a single `Dockerfile`. Thanks to multi-stage Dockerfiles, we don't loose any of our benefits that we did have with our `Dockerfile.builder` for example.

I've unified our Dockerfiles into one `Dockerfile` with three stages:

1. The first stage is called _build-tools_ and gives us a centos:7 image with all tools installed to build the software. No build or test is run in this stage.
2. The second stage, aka _builder_, is based upon the _build-tools_ stage and indeed builds the software and tests it (without e2e tests). 
3. The last and third stage, aka _deploy_, is again based on centos:7 and nothing else. None of the tools from _build_tools_ are available here. Then we copy over the build artifacts from the _builder_ stage to the _deploy_ stage.

## Artifacts

We nevery truely supported out-of-source builds and it is quite uncommon to do so with Go projects. That's why I've decided to simplify the makefiles a lot. You'll no longer find things like `$(CUR_DIR)/somefile`, but instead you'll see `./somefile`. The assumption is simply that you run `make` from the root directory of the project.

### ./out 

As a side-effect I decided to remove some more path variables as well. For example if your targets needs to write an artifact and location doesn't matter for you that much, you can safely assume that the directory `./out` is there. It will be cleaned up upon `make clean` for you, so you don't have to add a special clean target for your artifact.  

## Test simplification and cross-package coverage

As of go 1.9 `go list ./...` will exclude the vendor directory, so there's no need for us to filter out the vendor directory. Most of the test code that we had before was written by me in the github.com/fabric8-services/fabric8-wit repository and it was blindly copied over here without knowing what it does. I'm not particularly proud for the test code in fabric8-wit but back then it was the only way we could achieve _cross-package coverage_.

The idea for _cross-package coverage_ was to collect and merge the coverage information from package `a` and `b` if they call functions in tests from one another. Theoretically this sounds like a good idea. But consider that you want to take package `b` and use it somewhere else. Then all of a sudden the package coverage can drop drastically if most of the testing for `b` was done through package `a`.

I want a package to test it's code as best as it can and nothing more.

The result is what you can see in `./make/test.mk`. Testing the Go code can finally be fun again and if you want to take a look at the coverage of the project, simply run `make test-coverage` (for terminal coverage reports) or `make test-coverage-html` (for viewing line-by-line reports in your browser).  

## Linting YAML and Go

With the new `make lint-yaml` you can run `yamllint` on all non-vendor YAML files. 

I've made my life easy by simply `go get`ting the golang-ci instead of downloading a special build for each architecture and checking it's checksum before installing it. Once Go 1.13 is there we can `go get` a specific version.

## Finding tools

Instead of having a `prebuild` target in our makefile we're now looking for all the required tools on each invocation of `make`. The only exception is when you run `make help` or `make docker-XXX`. For `make help` we check for no tool at all. For `make docker-XXX` we just check for the `docker` executable. See the `./make/find-tools.mk` file for more information.

**NOTE:** We don't control which version each tool has as the moment.

## Verbosity

I wanted to be able to configure the output of `make` in these ways:

1. `make VERBOSE=0 <TARGET>`: Reduce output to minimum (not even showing the command that is executed).
2. `make VERBOSE=1 <TARGET>` (the default): Show the command that is executed but don't turn on `-v` verbosity flags for each command.
3. `make VERBOSE=2 <TARGET>`: Show the command as in 2. but also turn on all verbosity switches and disable silent flags like `-s`.

Have a look at the usage of variables like `${V_FLAG}` and `${S_FLAG}` to find out how to use them.

By prefixing a command with `$(Q)` instead of `@` we can control if the command shall be printed before executing. `$(Q)` expands to `@` when `make` is called with `VERBOSE=0` and to `` (empty) when called with any other number.