feat(ODC-451): add GitSource controller that checks if repo is reachable and branch exists (#118)
Adds GitSource controller that using a function from [devconsole-git](https://github.com/redhat-developer/devconsole-git/pull/23) checks if the repo is reachable or not. If it is not, then it means that either URL is wrong or the repo is private. In this case the operator sets:
*  `connection.state` to `failed` and `connection.error` to `unable to reach the repo`. 

If the repo is reachable then it checks if the provided branch exists. If it doesn't then it sets:
*  `connection.state` to `failed` and `connection.error` to `unable to find the branch`. 

If the repo is reachable and branch exists then it sets:
*  `connection.state` to `ok` and `connection.error` sets to empty string

this PR depends on:
https://github.com/redhat-developer/devconsole-git/pull/23 and https://github.com/redhat-developer/devconsole-operator/pull/91