Test operator source (#181)
This PR:

* Adds a make target (`test-operator-source`) to execute a simple test for verifying that the operator is successfully installed from application registry on Quay.io.

It is a functional test for `OperatorSource`. The files included in the pull request define the `OperatorSource` itself and it's corresponding `CatalogSource`, and `Subscription`.

It depends on the `push-operator-app-registry` make target from https://github.com/redhat-developer/devconsole-operator/pull/174 that creates a CSV for the operator and pushes it to quay.io. The `test-operator-source` make target runs the test to verify the creation of the `Subscription`, `InstallPlan`, and verifies that the operator's pod has started.